Removed styles to another file and some debugging code
